Moving svgs to new place where storybook can use it, updated storybook to simpler scripts for easier reading, renaming all variables that used the wrong location to avoid breaking the project
