+++ components/media-cta.scss
@@ -7,7 +7,7 @@
     @include card();
     text-align: center;
     width: 520px;
-    background-image: url('/svgs/assets/person.svg');
+    background-image: url('/static/svgs/assets/person.svg');
     background-repeat: no-repeat;
     background-position: center;
     background-size: 140px;
