+++ components/media-highlights.scss
@@ -28,11 +28,11 @@
   }
 
   .MyBitDappIcon:before {
-    background: url('/svgs/assets/mybit-dapp-icon.svg') no-repeat center;
+    background: url('/static/svgs/assets/mybit-dapp-icon.svg') no-repeat center;
   }
 
   .MyDaxIcon:before {
-    background: url('/svgs/assets/mydax-icon.svg') no-repeat center;
+    background: url('/static/svgs/assets/mydax-icon.svg') no-repeat center;
     background-size: 110px;
   }
