+++ components/wallets.scss
@@ -21,32 +21,32 @@
   }
 
   &--android {
-    background-image: url('/svgs/assets/android.svg'), url('/svgs/assets/mobile.svg');
+    background-image: url('/static/svgs/assets/android.svg'), url('/static/svgs/assets/mobile.svg');
     background-position: 50% 20%, top;
     background-size: 50px, 80px;
   }
 
   &--ios {
-    background: url('/svgs/assets/apple.svg'), url('/svgs/assets/mobile.svg');
+    background: url('/static/svgs/assets/apple.svg'), url('/static/svgs/assets/mobile.svg');
     background-position: top;
     background-position: 50% 20%, top;
     background-size: 50px, 80px;
   }
 
   &--mac {
-    background: url('/svgs/assets/apple.svg'), url('/svgs/assets/computer.svg');
+    background: url('/static/svgs/assets/apple.svg'), url('/static/svgs/assets/computer.svg');
     background-position: 50% 20%, top;
     background-size: 50px, 210px;
   }
 
   &--windows {
-    background: url('/svgs/assets/windows.svg'), url('/svgs/assets/computer.svg');
+    background: url('/static/svgs/assets/windows.svg'), url('/static/svgs/assets/computer.svg');
     background-position: 50% 20%, top;
     background-size: 55px, 210px;
   }
 
   &--myetherwallet {
-    background: url('/svgs/assets/myetherwallet.svg'), url('/svgs/assets/computer.svg');
+    background: url('/static/svgs/assets/myetherwallet.svg'), url('/static/svgs/assets/computer.svg');
     background-position: 50% 20%, top;
     background-size: 60px, 210px;
   }
