+++ package.json
@@ -91,9 +91,10 @@
     "build": "next build && next export -o build",
     "start": "NODE_ENV=production node server.js",
     "storybook": "start-storybook -s ./ -p 6006",
-    "clean-storybook": "mv .out/fonts .out/static/fonts && mv .out/assets .out/static/assets && del .out/mybit.{git,mp4} && del .out/{favicons,files,fonts,industries,news,offers,partners,social,team}",
+    "move-static": "mv .out/fonts .out/static/fonts && mv .out/assets .out/static/assets && mv .out/svgs .out/static/svgs",
+    "clean-storybook": "del .out/mybit.{git,mp4} && del .out/{favicons,files,fonts,industries,news,offers,partners,social,team}",
     "static-storybook": "build-storybook -s ./static -c .storybook -o .out",
-    "build-storybook": "npm-run-all -s static-storybook clean-storybook",
+    "build-storybook": "npm-run-all -s static-storybook move-static clean-storybook",
     "deploy-storybook:dev": "surge .out/ --domain storybook-mybit-io-dev.surge.sh",
     "deploy-storybook:prod": "surge .out/ --domain storybook-mybit-io.surge.sh",
     "deploy-storybook": "surge .out/",
